# docs/swagger.yaml
basePath: /api
definitions:
  challenge.CreateChallengeRequest:
    properties:
      assigneeIds:
        items:
          type: string
        type: array
      description:
        type: string
      expiresAt:
        type: string
      points:
        type: integer
      title:
        type: string
      type:
        type: integer
    required:
    - points
    - title
    type: object
...
  models.AssignmentStatus:
    enum:
    - 0
    - 1
    - 2
    type: integer
    x-enum-varnames:
    - AssignmentAssigned
    - AssignmentSubmitted
    - AssignmentApproved
  models.Challenge:
    properties:
      createdAt:
        type: string
      creatorId:
        type: string
      description:
        type: string
      expiresAt:
        type: string
      id:
        type: string
      points:
        type: integer
      restricted:
        type: boolean
      status:
        $ref: '#/definitions/models.ChallengeStatus'
      title:
        type: string
      type:
        $ref: '#/definitions/models.ChallengeType'
      updatedAt:
        type: string
    type: object
  models.ChallengeAssignment:
    properties:
      approvedAt:
        type: string
      assigneeId:
        type: string
      challengeId:
        type: string
      id:
        type: string
      status:
        $ref: '#/definitions/models.AssignmentStatus'
      submittedAt:
        type: string
    type: object
  models.ChallengeStatus:
    enum:
    - 0
    - 1
    - 2
    type: integer
    x-enum-varnames:
    - StatusActive
    - StatusExpired
    - StatusCancelled
  models.ChallengeType:
    enum:
    - 0
    - 1
    - 2
    type: integer
    x-enum-varnames:
    - Bounty
    - Daily
    - Weekly
...
  models.Message:
    properties:
      body:
        type: string
      createdAt:
        type: string
      id:
        type: string
      readAt:
        type: string
      recipientId:
        type: string
      senderId:
        type: string
    type: object
...
info:
  contact: {}
  description: A health and longevity challenge API
  title: Panjang Umur API
  version: "1.0"
paths:
...
  /challenges:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Challenge'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List active challenges visible to me (created by me, assigned to me,
        or open challenges from my friends)
      tags:
      - Challenges
    post:
      consumes:
      - application/json
      parameters:
      - description: Challenge data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/challenge.CreateChallengeRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Challenge'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create a challenge for one or more friends
      tags:
      - Challenges
  /challenges/{challengeId}/cancel:
    patch:
      parameters:
      - description: Challenge ID (UUID)
        in: path
        name: challengeId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Challenge'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Cancel an active challenge
      tags:
      - Challenges
  /challenges/{challengeId}/submit:
    patch:
      parameters:
      - description: Challenge ID (UUID)
        in: path
        name: challengeId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Challenge'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Submit a challenge for approval
      tags:
      - Challenges
  /challenges/{id}:
    get:
      parameters:
      - description: Challenge ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Challenge'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get a challenge by ID
      tags:
      - Challenges
  /challenges/{id}/assignments:
    get:
      parameters:
      - description: Challenge ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.ChallengeAssignment'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List a challenge's per-assignee assignments (submit/approve status)
      tags:
      - Challenges
  /challenges/assignments/{assignmentId}/approve:
    patch:
      parameters:
      - description: Assignment ID (UUID)
        in: path
        name: assignmentId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ChallengeAssignment'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Approve a submitted assignment and award points to that assignee
      tags:
      - Challenges
  /challenges/me:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Challenge'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get challenges assigned to the current user that are still open or
        awaiting approval
      tags:
      - Challenges
  /challenges/me/created:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Challenge'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get challenges created by the current user
      tags:
      - Challenges
  /chat/{friendId}/messages:
    get:
      parameters:
      - description: Friend's user ID (UUID)
        in: path
        name: friendId
        required: true
        type: string
      - description: Only return messages older than this RFC3339 timestamp (pagination
          cursor)
        in: query
        name: before
        type: string
      - description: Max messages to return (default 50, capped at 100)
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Message'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get message history with a friend
      tags:
      - Chat
  /chat/{friendId}/read:
    post:
      parameters:
      - description: Friend's user ID (UUID)
        in: path
        name: friendId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Mark every message sent by a friend to the caller as read
      tags:
      - Chat
...
securityDefinitions:
  BearerAuth:
    description: Use "Bearer <token>"
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"